chore: interval subscription manager context (#598)
